bug: ex00에서 음수의 출력 수정
int 범위 안임에도 출력이 impossible로 나오는 문제가 있었다.
numeric_limits<double>을 int로 수정하여 해결했다.

+++ ex00/ScalarConverter.cpp
@@ -79,7 +79,7 @@ void ScalarConverter::stringToOtherType(std::string& target)
 		std::cerr << RED << e.what() << RESET << '\n';
 		return ;
 	}
-	if (_doubleType != _doubleType || _doubleType > std::numeric_limits<double>::max() || _doubleType < std::numeric_limits<double>::min())
+	if (_doubleType != _doubleType || _doubleType > std::numeric_limits<int>::max() || _doubleType < std::numeric_limits<int>::min())
 	{
 		std::cout << "char: impossible" << std::endl;
 		std::cout << "int: impossible" << std::endl;
